# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pollen",
    "description": "Public API for the Tamanu deployment onboarding wizard.",
    "contact": {
      "name": "BES Developers",
      "email": "contact@bes.au"
    },
    "license": {
      "name": "GPL-3.0-or-later"
    },
    "version": ""
  },
  "paths": {
    "/api/applications/create": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create a new draft, binding the bundled default ruleset or a previewed branch.",
        "operationId": "applications_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppView"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/finalise": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Finalise a draft, freezing it against its bound ruleset. Always produces an\nartifact; the verdict (including a blocking one) is recorded, not enforced.",
        "operationId": "applications_finalise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinaliseArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppView"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/fork": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Fork a new draft from any application, with lineage to it. `to_default`\nrebinds to the bundled default; a named branch rebinds to that previewed\nruleset; otherwise the parent's ruleset is kept. Rebinding migrates the\nanswers (stable-id set-diff). The parent is left untouched.",
        "operationId": "applications_fork",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppView"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/get": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Read an application with its current evaluation.",
        "operationId": "applications_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppView"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/patch": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Replace a draft's answers. Rejected (409) once finalised.",
        "operationId": "applications_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppView"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/version": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Service name and version. Doubles as a trivial reachability check.",
        "operationId": "meta_version",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionInfo"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppView": {
        "type": "object",
        "description": "The full state of an application: its lifecycle fields, the questions to\nrender, the current answers, the evaluation (consequences, verdict, derived\nvalues, visible questions, guidance), and, on a fork, what changed.",
        "required": [
          "id",
          "status",
          "created_at",
          "config_hash",
          "update_available",
          "sections",
          "questions",
          "answers",
          "evaluation"
        ],
        "properties": {
          "answers": {},
          "config_hash": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "evaluation": {
            "$ref": "#/components/schemas/Evaluation"
          },
          "finalised_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "migration": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MigrationView"
              }
            ]
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "description": "The flow's sections, in presentation order."
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "update_available": {
            "type": "boolean",
            "description": "True when the plan is bound to a ruleset other than the current bundled\ndefault, meaning a newer default is available. A draft updates in place; a\nfinalised plan spawns a new draft on the new ruleset."
          }
        }
      },
      "ApplicationStatus": {
        "type": "string",
        "description": "Whether an artifact is still being edited or has been frozen.",
        "enum": [
          "draft",
          "finalised"
        ]
      },
      "Assumed": {
        "type": "object",
        "description": "A default the engine applied because the question was left unanswered.",
        "required": [
          "question",
          "option"
        ],
        "properties": {
          "option": {
            "type": "string"
          },
          "question": {
            "type": "string"
          }
        }
      },
      "Audience": {
        "type": "string",
        "description": "Which reader a consequence is grouped under in the by-audience view.",
        "enum": [
          "Client",
          "Bes",
          "Record"
        ]
      },
      "Consequence": {
        "type": "object",
        "required": [
          "status",
          "audience",
          "title",
          "detail"
        ],
        "properties": {
          "audience": {
            "$ref": "#/components/schemas/Audience"
          },
          "cost": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Cost"
              }
            ]
          },
          "detail": {
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "title": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsequenceType"
            }
          }
        }
      },
      "ConsequenceType": {
        "type": "string",
        "description": "The \"this is worse\" axis (spec WIZ, Consequence type).",
        "enum": [
          "Cost",
          "Operational",
          "Capability",
          "Support"
        ]
      },
      "Cost": {
        "type": "object",
        "required": [
          "tier"
        ],
        "properties": {
          "ballpark": {
            "type": [
              "string",
              "null"
            ]
          },
          "tier": {
            "type": "string"
          }
        }
      },
      "CreateArgs": {
        "type": "object",
        "properties": {
          "config_branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "A ruleset branch to preview; omitted binds the bundled default."
          }
        }
      },
      "Evaluation": {
        "type": "object",
        "required": [
          "derived",
          "visible_questions",
          "consequences",
          "guidance",
          "assumed",
          "open_items",
          "required",
          "verdict"
        ],
        "properties": {
          "assumed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Assumed"
            },
            "description": "Visible questions left unanswered whose blessed-path default the engine\napplied on the user's behalf (spec WIZ, assumed defaults)."
          },
          "consequences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriggeredConsequence"
            },
            "description": "Every triggered consequence, in ruleset order."
          },
          "derived": {
            "type": "object",
            "description": "Derived values keyed by derivation id (e.g. `size` → `Medium`).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "guidance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriggeredGuidance"
            },
            "description": "Guidance whose condition currently holds."
          },
          "open_items": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Visible questions the user marked unsure, or left blank where an unsure\noption was available. These make the artifact interim."
          },
          "required": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Visible questions that must be answered: no default to fall back on and\nno unsure option to decline with. These block finalising."
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "visible_questions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The ids of questions currently shown, in ruleset order (spec WIZ,\nvisibility). The frontend renders exactly these."
          }
        }
      },
      "FinaliseArgs": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ForkArgs": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "config_branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "A ruleset branch to rebind to; omitted keeps the parent's ruleset."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "to_default": {
            "type": "boolean",
            "description": "Rebind to the current bundled default ruleset (takes precedence over\n`config_branch`). Used to update a stale draft to the latest default."
          }
        }
      },
      "GetArgs": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MigrationView": {
        "type": "object",
        "description": "What a fork's migration changed (spec WIZ, stable-id migration).",
        "required": [
          "dropped",
          "new_questions"
        ],
        "properties": {
          "dropped": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "new_questions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Opt": {
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "properties": {
          "exclusive": {
            "type": "boolean",
            "description": "In a multi-select, choosing this option clears the others, and choosing\nany other clears this one (e.g. a \"none of these\" choice)."
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "unsure": {
            "type": "boolean",
            "description": "This option means \"I don't know\". Choosing it assumes nothing and\nrecords the question as an open item (spec WIZ, open items)."
          },
          "warn": {
            "type": [
              "string",
              "null"
            ],
            "description": "Shown inline when this option is selected, so the cost of leaving the\nblessed path lands at the point of choice."
          }
        }
      },
      "PatchArgs": {
        "type": "object",
        "required": [
          "id",
          "answers"
        ],
        "properties": {
          "answers": {
            "description": "Answers keyed by question id; each value is an option id or a list of them."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ProblemDetailsSchema": {
        "type": "object",
        "description": "Wire-shape mirror of [`problem_details::ProblemDetails`] for OpenAPI.\n\nEvery error response serializes to this RFC 7807 shape via [`AppError`]'s\n`IntoResponse`. This struct exists so OpenAPI specs can reference a concrete\nschema; nothing constructs values of it at runtime.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code generated by the origin server.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary of the problem."
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference identifying the problem type.",
            "example": "/errors/not-found"
          }
        }
      },
      "QuestionKind": {
        "type": "string",
        "enum": [
          "Single",
          "Multi",
          "Band"
        ]
      },
      "QuestionView": {
        "type": "object",
        "description": "A question's render metadata (the engine decides visibility; see\n`Evaluation::visible_questions`).",
        "required": [
          "id",
          "kind",
          "label",
          "options"
        ],
        "properties": {
          "default": {
            "type": [
              "string",
              "null"
            ],
            "description": "The option assumed when this is left blank, if any."
          },
          "help": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/QuestionKind"
          },
          "label": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Opt"
            }
          },
          "section": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Section": {
        "type": "object",
        "description": "A named group of questions, presented together.",
        "required": [
          "id",
          "label"
        ],
        "properties": {
          "collapsed": {
            "type": "boolean",
            "description": "Collapsed on arrival, for sections a non-technical user can skip."
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "Severity": {
        "type": "string",
        "description": "The viability axis (spec WIZ, Severity).",
        "enum": [
          "Default",
          "NonDefault",
          "Blocking"
        ]
      },
      "Status": {
        "type": "string",
        "description": "The technical-versus-contractual line (spec WIZ, Status).",
        "enum": [
          "Requirement",
          "Advisory",
          "Referral"
        ]
      },
      "TriggeredConsequence": {
        "type": "object",
        "required": [
          "id",
          "source",
          "consequence"
        ],
        "properties": {
          "consequence": {
            "$ref": "#/components/schemas/Consequence"
          },
          "id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
      "TriggeredGuidance": {
        "type": "object",
        "required": [
          "at",
          "message"
        ],
        "properties": {
          "at": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Verdict": {
        "type": "string",
        "description": "The viability verdict: the worst severity present across triggered\nconsequences. Default-severity consequences don't move it off `Clear`.",
        "enum": [
          "Clear",
          "NonDefault",
          "Blocking"
        ]
      },
      "VersionInfo": {
        "type": "object",
        "required": [
          "name",
          "version"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "applications",
      "description": "Onboarding artifacts: create, read, edit, finalise, fork."
    },
    {
      "name": "meta",
      "description": "Service metadata."
    }
  ]
}